+++ src/garage/np/algos/__init__.py
@@ -4,5 +4,8 @@ from garage.np.algos.batch_polopt import BatchPolopt
 from garage.np.algos.cem import CEM
 from garage.np.algos.cma_es import CMAES
 from garage.np.algos.nop import NOP
+from garage.np.algos.off_policy_rl_algorithm import OffPolicyRLAlgorithm
 
-__all__ = ['RLAlgorithm', 'BatchPolopt', 'CEM', 'CMAES', 'NOP']
+__all__ = [
+    'RLAlgorithm', 'BatchPolopt', 'CEM', 'CMAES', 'NOP', 'OffPolicyRLAlgorithm'
+]
