+++ src/garage/tf/algos/__init__.py
@@ -3,7 +3,6 @@ from garage.tf.algos.ddpg import DDPG
 from garage.tf.algos.dqn import DQN
 from garage.tf.algos.erwr import ERWR
 from garage.tf.algos.npo import NPO
-from garage.tf.algos.off_policy_rl_algorithm import OffPolicyRLAlgorithm
 from garage.tf.algos.ppo import PPO
 from garage.tf.algos.reps import REPS
 from garage.tf.algos.tnpg import TNPG
@@ -16,7 +15,6 @@ __all__ = [
     'DQN',
     'ERWR',
     'NPO',
-    'OffPolicyRLAlgorithm',
     'PPO',
     'REPS',
     'TNPG',
