+++ tests/fixtures/tf/algos/dummy_off_policy_algo.py
@@ -1,4 +1,4 @@
-from garage.tf.algos import OffPolicyRLAlgorithm
+from garage.np.algos import OffPolicyRLAlgorithm
 
 
 class DummyOffPolicyAlgo(OffPolicyRLAlgorithm):
